+++ kbp/server.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD
 // license that can be found in the LICENSE file.
 
-package libpages
+package kbp
 
 import (
 	"context"
